pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psi"
version = "1.0.0"
description = "FastAPI Application with Advanced Logging"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "sqlmodel",
    "alembic",
    "loguru",
    "pretty-errors",
    "rich",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "redis",
    "prometheus-client",
    "prometheus-fastapi-instrumentator",
    "pydantic-settings",
    "python-dotenv",
    "debugpy",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py38"
